Refactor code and check style
